feat: implement authentication context and session management
