+++ app/(app)/(stacks)/planDetail.tsx
