+++ app/+not-found.tsx
@@ -3,14 +3,14 @@ import { StyleSheet, View } from 'react-native';
 
 export default function NotFoundScreen() {
   return (
-    <>
+    <View>
       <Stack.Screen options={{ title: 'Oops! Not Found' }} />
       <View style={styles.container}>
         <Link href="/" style={styles.button}>
           Go back to Home screen!
         </Link>
       </View>
-    </>
+    </View>
   );
 }
