+++ ctx.tsx
@@ -1,0 +1,41 @@
+import { createContext, useContext, type PropsWithChildren } from 'react';
+import { useStorageState } from './useStorageState';
+
+const AuthContext = createContext<{
+  signIn: () => void;
+  signOut: () => void;
+  session?: string | null;
+  isLoading: boolean;
+}>({
+  signIn: () => null,
+  signOut: () => null,
+  session: null,
+  isLoading: false,
+});
+
+export function useSession() {
+  const value = useContext(AuthContext);
+
+  return value;
+}
+
+export function SessionProvider({ children }: PropsWithChildren) {
+  const [[isLoading, session], setSession] = useStorageState('session');
+
+  return (
+    <AuthContext.Provider
+      value={{
+        signIn: async () => {
+          // Perform sign-in logic here
+          setSession('xxx');
+        },
+        signOut: () => {
+          setSession(null);
+        },
+        session,
+        isLoading,
+      }}>
+      {children}
+    </AuthContext.Provider>
+  );
+}
